Allow stylesheet selection
The XSLT_DIR environment variable is now used to set the directory
with stylesheets, current directory is used by default.

This is intended to be combined with web server
configuration (rewrites in particular) and "hosts" file or local DNS
server records, in order to mimic other services, and to use this
API/DB without adjusting clients.