Use Makefile.PL
Now ExtUtils::MakeMaker is used to provide relatively nice
installation. Though `make uninstall` is broken there, and maybe it
should be replaced with autotools or a different module.

File::ShareDir::Install and File::ShareDir are used to install files
into a data directory and locate it. The latter doesn't seem to handle
INSTALL_BASE.